[chore] renaming in camlCase
