+++ curator/src/mail_agent/newsletterTxtFormat.ts
